pyproject.toml
[tool.poetry]
name = "blender-studio"
version = "0.1.0"
description = ""
authors = ["Sem Mulder <sem@blender.studio>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2"
django-pipeline = "^2.0"
psycopg2 = "^2.8"
blender_id_oauth_client = {git = "https://gitlab.com/blender-institute/blender-id-oauth-client.git", branch = "master"}
libsasscompiler = "^0.1.5"
jsmin = "3.0.0"
sorl-thumbnail = "^12.7.0"
mistune = "2.0.0a4"
looper = {git = "https://projects.blender.org/infrastructure/looper.git", branch = "master"}
Pillow = "^8.0"
django-storages = {extras = ["google"], version = "1.11.1"}
pymongo = "^3.10.1"
markupsafe = "^1.1.1"
meilisearch = "^0.18.0"
django-taggit = "^1.3.0"
boto3 = "1.18.56"
responses = "^0.12.0"
attrs = "^19.3.0"
bleach = "^3.2.1"
shortcodes = "2.5"
sentry-sdk = "^1.4.3"
requests-oauthlib = "^1.3.0"
django-activity-stream = "^0.9.0"
django-background-tasks = "^1.2.5"
django-anymail = {extras = ["mailgun"], version = "8.2"}
django-nested-admin = "^3.3.3"
html5lib = "1.1"
braintree = "4.17.1"
python-stdnum = "^1.16"
localflavor = "^1.9"
pyvat = {git = "https://github.com/iconfinder/pyvat.git", rev = "419abd659ae5a4a6cb6ea9b54aa4bde17aefeb5b"}
alphabetic-timestamp = "1.1.5"
zeep = "^4.0.0"
django-loginas = "^0.3.9"
Markdown = "^3.3.6"
djangorestframework = "^3.13.1"
tldextract = "^3.3.1"
dj-database-url = "1.0.0"
python-dotenv = "^0.21.0"
reportlab = "3.6.6"  # see https://github.com/xhtml2pdf/xhtml2pdf/pull/590
django-modeltranslation = "^0.18.11"
stripe = "^6.1.0"
googletrans = "3.1.0a0"
flask = "1.1.4"
django-import-export = "^3.2.0"

[tool.poetry.dev-dependencies]
black = "21.6b0"
mypy = "*"
django-stubs = "^1.5"
pre-commit = "2.16.0"
ipython = "^7.17"
factory-boy = "^3.0"
django-debug-toolbar = "^2.2"
flake8 = "^3.8.3"
flake8-docstrings = "^1.5.0"
freezegun = "^1.0.0"
django-sslserver = "^0.22"
djhtml = "1.4.0"
phpserialize = "^1.3"
PyPDF2 = "^1.26.0"

[tool.black]
line-length = 100
target_version = ['py38']
skip_string_normalization = true
exclude = 'migrations/|studio/settings'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
